Moved ExcelComAddInHelper to a separate file.
